refactor: Moved PlaybackControl into a separate component
